bevy_input::gamepad
Enum GamepadRumbleRequest

pub enum GamepadRumbleRequest {
    Add {
        duration: Duration,
        intensity: GamepadRumbleIntensity,
        gamepad: Entity,
    },
    Stop {
        gamepad: Entity,
    },
}

An event that controls force-feedback rumbling of a Gamepad entity.

Notes

Does nothing if the gamepad or platform does not support rumble.

Example

fn rumble_gamepad_system(
    mut rumble_requests: EventWriter<GamepadRumbleRequest>,
    gamepads: Query<Entity, With<Gamepad>>,
) {
    for entity in gamepads.iter() {
        rumble_requests.send(GamepadRumbleRequest::Add {
            gamepad: entity,
            intensity: GamepadRumbleIntensity::MAX,
            duration: Duration::from_secs_f32(0.5),
        });
    }
}

Variants

Add

Add a rumble to the given gamepad.

Simultaneous rumble effects add up to the sum of their strengths.

Consequently, if two rumbles at half intensity are added at the same time, their intensities will be added up, and the controller will rumble at full intensity until one of the rumbles finishes, then the rumble will continue at the intensity of the remaining event.

To replace an existing rumble, send a GamepadRumbleRequest::Stop event first.

Fields

duration: Duration

How long the gamepad should rumble.

intensity: GamepadRumbleIntensity

How intense the rumble should be.

gamepad: Entity

The gamepad to rumble.

Stop

Stop all running rumbles on the given Entity.

Fields

gamepad: Entity

The gamepad to stop rumble.

Implementations

impl GamepadRumbleRequest

pub fn gamepad(&self) -> Entity

Get the Entity associated with this request.