bevy_input
Module mouse

The mouse input functionality.

Structs

AccumulatedMouseMotion

Tracks how much the mouse has moved every frame.

AccumulatedMouseScroll

Tracks how much the mouse has scrolled every frame.

MouseButtonInput

A mouse button input event.

MouseMotion

An event reporting the change in physical position of a pointing device.

MouseWheel

A mouse wheel event.

Enums

MouseButton

A button on a mouse device.

MouseScrollUnit

The scroll unit.

Functions

accumulate_mouse_motion_system

Updates the AccumulatedMouseMotion resource using the MouseMotion event. The value of AccumulatedMouseMotion is reset to zero every frame

accumulate_mouse_scroll_system

Updates the AccumulatedMouseScroll resource using the MouseWheel event. The value of AccumulatedMouseScroll is reset to zero every frame

mouse_button_input_system

Updates the ButtonInput<MouseButton> resource with the latest MouseButtonInput events.