Enum bevy_input::gamepad::GamepadRumbleRequest

pub enum GamepadRumbleRequest {
    Add {
        duration: Duration,
        intensity: GamepadRumbleIntensity,
        gamepad: Gamepad,
    },
    Stop {
        gamepad: Gamepad,
    },
}

An event that controls force-feedback rumbling of a Gamepad

Notes

Does nothing if the gamepad or platform does not support rumble.

Example

fn rumble_gamepad_system(
    mut rumble_requests: EventWriter<GamepadRumbleRequest>,
    gamepads: Res<Gamepads>
) {
    for gamepad in gamepads.iter() {
        rumble_requests.send(GamepadRumbleRequest::Add {
            gamepad,
            intensity: GamepadRumbleIntensity::MAX,
            duration: Duration::from_secs_f32(0.5),
        });
    }
}

Variants

Add

Fields

duration: Duration

How long the gamepad should rumble

intensity: GamepadRumbleIntensity

How intense the rumble should be

gamepad: Gamepad

The gamepad to rumble

Add a rumble to the given gamepad.

Simultaneous rumble effects add up to the sum of their strengths.

Consequently, if two rumbles at half intensity are added at the same time, their intensities will be added up, and the controller will rumble at full intensity until one of the rumbles finishes, then the rumble will continue at the intensity of the remaining event.

To replace an existing rumble, send a GamepadRumbleRequest::Stop event first.

Stop

Fields

gamepad: Gamepad

Stop all running rumbles on the given Gamepad

Implementations

impl GamepadRumbleRequest

pub fn gamepad(&self) -> Gamepad

Get the Gamepad associated with this request